client-go/util/consistencydetector: move IsDataConsistencyDetectionForWatchListEnabled
